Moved buttons and logic to components
